+++ framework/web/Request.php
@@ -94,10 +94,10 @@ class Request extends \yii\base\Request
 	 * from the same application. If not, a 400 HTTP exception will be raised.
 	 *
 	 * Note, this feature requires that the user client accepts cookie. Also, to use this feature,
-	 * forms submitted via POST method must contain a hidden input whose name is specified by [[csrfVar]].
+	 * forms submitted via POST method must contain a hidden input whose name is specified by [[csrfParam]].
 	 * You may use [[\yii\web\Html::beginForm()]] to generate his hidden input.
 	 *
-	 * In JavaScript, you may get the values of [[csrfVar]] and [[csrfToken]] via `yii.getCsrfParam()` and
+	 * In JavaScript, you may get the values of [[csrfParam]] and [[csrfToken]] via `yii.getCsrfParam()` and
 	 * `yii.getCsrfToken()`, respectively. The [[\yii\web\YiiAsset]] asset must be registered.
 	 *
 	 * @see Controller::enableCsrfValidation
@@ -108,7 +108,7 @@ class Request extends \yii\base\Request
 	 * @var string the name of the token used to prevent CSRF. Defaults to '_csrf'.
 	 * This property is used only when [[enableCsrfValidation]] is true.
 	 */
-	public $csrfVar = '_csrf';
+	public $csrfParam = '_csrf';
 	/**
 	 * @var array the configuration of the CSRF cookie. This property is used only when [[enableCsrfValidation]] is true.
 	 * @see Cookie
@@ -124,7 +124,7 @@ class Request extends \yii\base\Request
 	 * @see getMethod()
 	 * @see getBodyParams()
 	 */
-	public $methodVar = '_method';
+	public $methodParam = '_method';
 	/**
 	 * @var array the parsers for converting the raw HTTP request body into [[bodyParams]].
 	 * The array keys are the request `Content-Types`, and the array values are the
@@ -207,8 +207,8 @@ class Request extends \yii\base\Request
 	 */
 	public function getMethod()
 	{
-		if (isset($_POST[$this->methodVar])) {
-			return strtoupper($_POST[$this->methodVar]);
+		if (isset($_POST[$this->methodParam])) {
+			return strtoupper($_POST[$this->methodParam]);
 		} elseif (isset($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE'])) {
 			return strtoupper($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE']);
 		} else {
@@ -330,9 +330,9 @@ class Request extends \yii\base\Request
 	{
 		if ($this->_bodyParams === null) {
 			$contentType = $this->getContentType();
-			if (isset($_POST[$this->methodVar])) {
+			if (isset($_POST[$this->methodParam])) {
 				$this->_bodyParams = $_POST;
-				unset($this->_bodyParams[$this->methodVar]);
+				unset($this->_bodyParams[$this->methodParam]);
 			} elseif (isset($this->parsers[$contentType])) {
 				$parser = Yii::createObject($this->parsers[$contentType]);
 				if (!($parser instanceof RequestParserInterface)) {
@@ -1102,7 +1102,7 @@ class Request extends \yii\base\Request
 	public function getRawCsrfToken()
 	{
 		if ($this->_csrfCookie === null) {
-			$this->_csrfCookie = $this->getCookies()->get($this->csrfVar);
+			$this->_csrfCookie = $this->getCookies()->get($this->csrfParam);
 			if ($this->_csrfCookie === null) {
 				$this->_csrfCookie = $this->createCsrfCookie();
 				Yii::$app->getResponse()->getCookies()->add($this->_csrfCookie);
@@ -1174,7 +1174,7 @@ class Request extends \yii\base\Request
 	protected function createCsrfCookie()
 	{
 		$options = $this->csrfCookie;
-		$options['name'] = $this->csrfVar;
+		$options['name'] = $this->csrfParam;
 		$options['value'] = Security::generateRandomKey();
 		return new Cookie($options);
 	}
@@ -1193,8 +1193,8 @@ class Request extends \yii\base\Request
 		if (!$this->enableCsrfValidation || in_array($method, ['GET', 'HEAD', 'OPTIONS'], true)) {
 			return true;
 		}
-		$trueToken = $this->getCookies()->getValue($this->csrfVar);
-		$token = $this->getBodyParam($this->csrfVar);
+		$trueToken = $this->getCookies()->getValue($this->csrfParam);
+		$token = $this->getBodyParam($this->csrfParam);
 		return $this->validateCsrfTokenInternal($token, $trueToken)
 			|| $this->validateCsrfTokenInternal($this->getCsrfTokenFromHeader(), $trueToken);
 	}